refactor: Simplify code reviewer to prevent complexity doom loop
- Replaced 3000+ lines of complex code with 465 lines that just work
- Removed enhanced analyzers, deduplication, complex parsing
- Clear output: file, line, issue, fix - no garbage
- Added COMPLEXITY WARNING to prevent future over-engineering
- Kept multiple prompt styles (default, brutal, lenient, security)
- Updated README with simplicity warning
- Reduced dependencies to just 'requests'

This is the THIRD refactor due to complexity creep. The simple version
works perfectly. DO NOT add complexity.
