+++ requirements.txt
@@ -1,3 +1,1 @@
-requests>=2.28.0
-pyyaml>=6.0
-pytest>=7.0.0
+requests>=2.31.0
